backend/requirements.txt
Django>=4.2
Pillow
djangorestframework
django-cors-headers
whitenoise
twilio